Added playStoneAnimation()
This function takes the stackHeight at which the stone in game logic got placed and plays an animation until it reaches the bottom the top of the current tower. If the tower is already full of stones nothing happens. This should later not be possible through game logic.

Also experimented with some IDLE/loading animations related to the playStoneAnimation which currently runs in the startup state.